Extracted extension method to Utils file
