+++ compose-destinations-ksp/src/main/kotlin/com/ramcosta/composedestinations/ksp/commons/Utils.kt
@@ -1,8 +1,11 @@
 package com.ramcosta.composedestinations.ksp.commons
 
+import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.*
 import com.ramcosta.composedestinations.codegen.commons.*
 import com.ramcosta.composedestinations.codegen.model.*
+import com.ramcosta.composedestinations.codegen.model.ClassKind
+import com.ramcosta.composedestinations.ksp.processors.KSPClassKind
 import java.io.*
 
 fun KSAnnotated.findAnnotation(name: String): KSAnnotation {
@@ -67,4 +70,48 @@ fun KSType.isRequireOptInAnnotation(): Boolean {
             annotation.annotationType.resolve().isRequireOptInAnnotation()
         }
     }
+}
+
+fun Resolver.getNavTypeSerializers(): List<NavTypeSerializer> {
+    return getSymbolsWithAnnotation(NAV_TYPE_SERIALIZER_ANNOTATION_QUALIFIED)
+        .filterIsInstance<KSClassDeclaration>().map {
+            if (it.classKind != KSPClassKind.CLASS && it.classKind != KSPClassKind.OBJECT) {
+                throw IllegalDestinationsSetup("${it.simpleName}: Type serializers must be either class or object!")
+            }
+
+            var superType: KSType? = null
+            var isCustomTypeSerializer = false
+            for (type in it.superTypes) {
+                val resolvedType = type.resolve()
+                val resolvedTypeString = resolvedType.declaration.qualifiedName?.asString()
+                if (resolvedTypeString ==
+                    "$CORE_PACKAGE_NAME.navargs.parcelable.ParcelableNavTypeSerializer") {
+                    superType = resolvedType
+                    break
+                }
+
+                if (resolvedTypeString ==
+                    "$CORE_PACKAGE_NAME.navargs.serializable.SerializableNavTypeSerializer") {
+                    superType = resolvedType
+                    break
+                }
+
+                if (resolvedTypeString == NAV_TYPE_SERIALIZER_CUSTOM_SERIALIZER) {
+                    superType = resolvedType
+                    isCustomTypeSerializer = true
+                    break
+                }
+            }
+            if (superType == null) {
+                throw IllegalDestinationsSetup("${it.simpleName}: Type serializers must implement ParcelableNavTypeSerializer (or SerializableNavTypeSerializer for Serializable types)!")
+            }
+            val genericType = superType.arguments.first().type?.resolve()?.declaration as KSClassDeclaration
+
+            NavTypeSerializer(
+                classKind = if (it.classKind == KSPClassKind.CLASS) ClassKind.CLASS else ClassKind.OBJECT,
+                serializerType = ClassType(it.simpleName.asString(), it.qualifiedName!!.asString()),
+                genericType = ClassType(genericType.simpleName.asString(), genericType.qualifiedName!!.asString()),
+                isCustomTypeSerializer = isCustomTypeSerializer,
+            )
+        }.toList()
 }
